Conflicts:
	cocos/2d/CCFastTMXLayer.cpp
	cocos/Android.mk
	cocos/audio/android/jni/cddandroidAndroidJavaEngine.cpp
	cocos/network/Android.mk
	cocos/platform/CCFileUtils.cpp
	cocos/platform/android/CCEnhanceAPI-android.cpp
	cocos/platform/android/CCFileUtils-android.cpp
	cocos/platform/android/java/project.properties
	cocos/scripting/lua-bindings/auto/lua_cocos2dx_spine_auto.hpp
	cocos/scripting/lua-bindings/manual/CCLuaStack.cpp
	cocos/scripting/lua-bindings/proj.android/Android.mk
	cocos/ui/UIVideoPlayer-ios.mm
	extensions/Android.mk
	external
	external/config.json
	tests/cpp-empty-test/CMakeLists.txt
	tests/cpp-empty-test/Classes/AppDelegate.cpp
	tests/cpp-tests/CMakeLists.txt
	tests/cpp-tests/Classes/BugsTest/Bug-PageViewLayout.cpp
	tests/cpp-tests/Classes/CocosStudio3DTest/CocosStudio3DTest.cpp
	tests/cpp-tests/Classes/ExtensionsTest/CocoStudioActionTimelineTest/ActionTimelineTestScene.cpp
	tests/cpp-tests/Classes/ExtensionsTest/CocoStudioActionTimelineTest/ActionTimelineTestScene.h
	tests/cpp-tests/Classes/ExtensionsTest/CocoStudioArmatureTest/ArmatureScene.cpp
	tests/cpp-tests/Classes/ExtensionsTest/CocoStudioArmatureTest/ArmatureScene.h
	tests/cpp-tests/Classes/ExtensionsTest/CocoStudioComponentsTest/EnemyController.h
	tests/cpp-tests/Classes/ExtensionsTest/CocoStudioComponentsTest/PlayerController.h
	tests/cpp-tests/Classes/ExtensionsTest/CocoStudioComponentsTest/ProjectileController.h
	tests/cpp-tests/Classes/ExtensionsTest/CocoStudioComponentsTest/SceneController.h
	tests/cpp-tests/Classes/ExtensionsTest/CocoStudioSceneTest/SceneEditorTest.cpp
	tests/cpp-tests/Classes/ExtensionsTest/CocoStudioSceneTest/SceneEditorTest.h
	tests/cpp-tests/Classes/ExtensionsTest/CocoStudioSceneTest/TriggerCode/acts.h
	tests/cpp-tests/Classes/ExtensionsTest/CocoStudioSceneTest/TriggerCode/cons.h
	tests/cpp-tests/Classes/ExtensionsTest/NetworkTest/WebSocketTest.cpp
	tests/cpp-tests/Classes/ExtensionsTest/NetworkTest/WebSocketTest.h
	tests/cpp-tests/Classes/LabelTest/LabelTestNew.cpp
	tests/cpp-tests/Classes/LabelTest/LabelTestNew.h
	tests/cpp-tests/Classes/NewRendererTest/NewRendererTest.cpp
	tests/cpp-tests/Classes/SpriteFrameCacheTest/SpriteFrameCacheTest.cpp
	tests/cpp-tests/Classes/SpritePolygonTest/SpritePolygonTest.cpp
	tests/cpp-tests/Classes/SpritePolygonTest/SpritePolygonTest.h
	tests/cpp-tests/Classes/UITest/CocoStudioGUITest/CocostudioParserTest/CocostudioParserJsonTest.cpp
	tests/cpp-tests/Classes/UITest/CocoStudioGUITest/CustomTest/CustomImageTest/CustomImageTest.cpp
	tests/cpp-tests/Classes/UITest/CocoStudioGUITest/CustomTest/CustomParticleWidgetTest/CustomParticleWidgetTest.cpp
	tests/cpp-tests/Classes/UITest/CocoStudioGUITest/CustomTest/CustomWidgetCallbackBindTest/CustomRootNode.h
	tests/cpp-tests/Classes/UITest/CocoStudioGUITest/CustomTest/CustomWidgetCallbackBindTest/CustomRootNodeReader.h
	tests/cpp-tests/Classes/UITest/CocoStudioGUITest/CustomTest/CustomWidgetCallbackBindTest/CustomWidgetCallbackBindTest.cpp
	tests/cpp-tests/Classes/UITest/CocoStudioGUITest/CustomWidget/CustomImageViewReader.h
	tests/cpp-tests/Classes/UITest/CocoStudioGUITest/CustomWidget/CustomParticleWidgetReader.cpp
	tests/cpp-tests/Classes/UITest/CocoStudioGUITest/CustomWidget/CustomParticleWidgetReader.h
	tests/cpp-tests/Classes/UITest/CocoStudioGUITest/CustomWidget/CustomReader.h
	tests/cpp-tests/Classes/UITest/CocoStudioGUITest/GUIEditorTest.cpp
	tests/cpp-tests/Classes/UITest/CocoStudioGUITest/UIListViewTest/UIListViewTest.cpp
	tests/cpp-tests/Classes/UITest/CocoStudioGUITest/UIPageViewTest/UIPageViewTest.cpp
	tests/cpp-tests/Classes/UITest/CocoStudioGUITest/UIRichTextTest/UIRichTextTest.cpp
	tests/cpp-tests/Classes/UITest/CocoStudioGUITest/UIRichTextTest/UIRichTextTest.h
	tests/cpp-tests/Classes/UITest/CocoStudioGUITest/UIScrollViewTest/UIScrollViewTest_Editor.cpp
	tests/cpp-tests/Classes/UITest/CocoStudioGUITest/UIScrollViewTest/UIScrollViewTest_Editor.h
	tests/cpp-tests/Classes/UITest/UITest.cpp
	tests/cpp-tests/Resources/ccs-res
	tests/cpp-tests/proj.android-studio/app/jni/Android.mk
	tests/cpp-tests/proj.android/jni/Android.mk
	tests/cpp-tests/proj.win10/Package.appxmanifest
	tests/cpp-tests/proj.win10/cpp-tests.vcxproj
	tests/cpp-tests/proj.win10/cpp-tests.vcxproj.filters
	tests/cpp-tests/proj.win32/cpp-tests.vcxproj
	tests/cpp-tests/proj.win32/cpp-tests.vcxproj.filters
	tests/cpp-tests/proj.win8.1-universal/cpp-tests.Shared/cpp-tests.Shared.vcxitems
	tests/cpp-tests/proj.win8.1-universal/cpp-tests.Shared/cpp-tests.Shared.vcxitems.filters
	tests/js-memory-gc-tests/project/Classes/AppDelegate.cpp
	tests/js-memory-gc-tests/project/Classes/AppDelegate.h
	tests/js-memory-gc-tests/project/proj.android-studio/app/AndroidManifest.xml
	tests/js-memory-gc-tests/project/proj.android-studio/app/jni/Android.mk
	tests/js-memory-gc-tests/project/proj.android-studio/app/jni/Application.mk
	tests/js-memory-gc-tests/project/proj.android-studio/app/jni/main.cpp
	tests/js-memory-gc-tests/project/proj.android/AndroidManifest.xml
	tests/js-memory-gc-tests/project/proj.android/jni/Android.mk
	tests/js-memory-gc-tests/project/proj.android/jni/Application.mk
	tests/js-memory-gc-tests/project/proj.android/jni/main.cpp
	tests/js-memory-gc-tests/project/proj.ios/AppController.h
	tests/js-memory-gc-tests/project/proj.ios/AppController.mm
	tests/js-memory-gc-tests/project/proj.ios/NativeOcClass.h
	tests/js-memory-gc-tests/project/proj.ios/NativeOcClass.m
	tests/js-tests/project/Classes/AppDelegate.cpp
	tests/js-tests/project/Classes/js_Effect3D_bindings.cpp
	tests/js-tests/src/CocoStudioTest/ArmatureTest/ArmatureTest.js
	tests/js-tests/src/CocosDenshionTest/CocosDenshionTest.js
	tests/js-tests/src/CocosNodeTest/CocosNodeTest.js
	tests/js-tests/src/GUITest/UIPageViewTest/UIPageViewTest.js
	tests/js-tests/src/GUITest/UISceneManager.js
	tests/js-tests/src/LoaderTest/LoaderTest.js
	tests/js-tests/src/SceneTest/SceneTest.js
	tests/js-tests/src/SpriteTest/SpriteTest.js
	tests/js-tests/src/tests_resources.js
	tests/lua-tests/project/Classes/lua_assetsmanager_test_sample.cpp
	tests/lua-tests/src/ByteCodeEncryptTest/ByteCodeEncryptTest.lua
	tests/lua-tests/src/CocoStudioTest/CocoStudioGUITest/CocoStudioGUITest.lua
	web

+++ cocos/scripting/js-bindings/manual/jsb_opengl_manual.cpp
@@ -1,6 +1,6 @@
 /*
  * Copyright (c) 2013 Zynga Inc.
- * Copyright (c) 2013-2014 Chukong Technologies Inc.
+ * Copyright (c) 2013-2016 Chukong Technologies Inc.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -380,7 +380,7 @@ bool JSB_glGetSupportedExtensions(JSContext *cx, uint32_t argc, jsval *vp)
     GLubyte* copy = new (std::nothrow) GLubyte[len+1];
     strncpy((char*)copy, (const char*)extensions, len );
 
-    int start_extension=0;
+    size_t start_extension = 0;
     int element=0;
     for( size_t i=0; i<len+1; i++) {
         if( copy[i]==' ' || copy[i]==',' || i==len ) {
@@ -519,7 +519,7 @@ bool JSB_glGetUniformfv(JSContext *cx, uint32_t argc, jsval *vp)
     JSObject *typedArray = NULL;
     if( utype == GL_FLOAT) {
         // FIXME: glew on windows will cause array overflow after invoking glGetUniformfv.
-        // It seems that glGetUniformfv re-assign the memeroy with a wrong size which is 4x than we pass in.
+        // It seems that glGetUniformfv re-assign the memory with a wrong size which is 4x than we pass in.
         // For temporary solution, we allocate 4x array. 
         GLfloat* param = new (std::nothrow) GLfloat[usize*4];
         glGetUniformfv(arg0, arg1, param);
@@ -530,7 +530,7 @@ bool JSB_glGetUniformfv(JSContext *cx, uint32_t argc, jsval *vp)
         CC_SAFE_DELETE_ARRAY(param);
     } else if( utype == GL_INT ) {
         // FIXME: glew on windows will cause array overflow after invoking glGetUniformfv.
-        // It seems that glGetUniformfv re-assign the memeroy with a wrong size which is 4x than we pass in.
+        // It seems that glGetUniformfv re-assign the memory with a wrong size which is 4x than we pass in.
         // For temporary solution, we allocate 4x array. 
         GLint* param = new (std::nothrow) GLint[usize*4];
         glGetUniformiv(arg0, arg1, param);
